Added support/fix for CS:GOs gamelump version 11 (uniformScale for props)
