+++ src/main/java/info/ata4/bspsrc/modules/entity/EntitySource.java
@@ -687,6 +687,10 @@ public class EntitySource extends ModuleDecompile {
                 }
             }
 
+            if (pst instanceof DStaticPropV11CSGO) {
+                writer.put("uniformscale", ((DStaticPropV11CSGO) pst).uniformScale);
+            }
+
             writer.end("entity");
         }
